requirements.txt
streamlit
smtplib
dnspython
pandas